+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 - Electra: build blocks with blobs.
 - E2E: fixed gas limit at genesis
 - Light client support: use LightClientHeader instead of BeaconBlockHeader.
+- Core: Fix process effective balance update to safe copy validator for Electra.
 
 ### Security
 
@@ -116,6 +117,7 @@ In an upcoming release, we will be deprecating the gRPC gateway and renaming sev
 - Fix Event stream with carriage return support
 - Fix panic on empty block result in REST API
 - engine_getPayloadBodiesByRangeV1 - fix, adding hexutil encoding on request parameters
+- Use sync committee period instead of epoch in `createLightClientUpdate` 
 
 
 ### Security